+++ src/main/java/kr/co/imguru/domain/like/repository/LikeReviewRepository.java
@@ -1,6 +1,6 @@
 package kr.co.imguru.domain.like.repository;
 
-import kr.co.imguru.domain.like.LikeReview;
+import kr.co.imguru.domain.like.entity.LikeReview;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
